cloud-backend/requirements.txt
# AutoQA Cloud Backend Dependencies

# Server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database (MongoDB)
pymongo>=4.0
dnspython>=2.6
certifi

# Validation / Config
pydantic-settings

# Auth
passlib[bcrypt]
python-jose[cryptography]
email-validator

# Environment
python-dotenv==1.0.1
